+++ src/functions/comparison/smaller.js
@@ -20,8 +20,8 @@ class LimeFunctionSmaller extends LimeFunction {
 
       'cl(var->expr)',
       'cr(var->expr)',
-      'cl(arg{expr}->expr)',
-      'cr(arg{expr}->expr)',
+      'cl(arg{expr[1]}->expr)',
+      'cr(arg{expr[1]}->expr)',
       'cl(bool->int)',
       'cr(bool->int)',
       'cb(int->rat,rat)',