+++ src/functions/counting/factorial.js
@@ -16,7 +16,7 @@ class LimeFunctionFactorial extends LimeFunction {
       'tf(!=)',
 
       'cl(var->expr)',
-      'cl(arg{expr}->expr)',
+      'cl(arg{expr[1]}->expr)',
       'cl(bool->int)',
 
       'el(int)',