+++ src/functions/logical/logicalNot.js
@@ -17,7 +17,7 @@ class LimeFunctionLogicalNot extends LimeFunction {
       'tr(_-)',
 
       'cr(var->expr)',
-      'cr(arg{expr}->expr)',
+      'cr(arg{expr[1]}->expr)',
       'cr(!bool->bool)',
 
       'er(bool)',