+++ src/functions/numeric/absolute.js
@@ -3,26 +3,25 @@ const LimeFunction = require('../../structs/function');
 
 /* ------------------------ division ------------------------ */
 
-// Fibonacci function class
-class LimeFunctionFibonacci extends LimeFunction {
+// Absolute function class
+class LimeFunctionAbsolute extends LimeFunction {
   // Constructor
   constructor(lime, mode) {
     // Super from function class
-    super(lime, { name: 'fibonacci', mode });
+    super(lime, { name: 'absolute', mode });
 
     // Operations
     this.operations.r = [
       'er(arg{int[1]})',
+      'er(arg{rat[1]})',
     ];
 
     // Algorithms
     this.algorithms.set('r(arg{int[1]})', (step) => {
-      const t = step.right.value[0];
-      if (t < 1) {
-        throw new Error('error:invalidArgumentLogic');
-      }
-
-      step.rus(this.lime.direct(t > 2 ? ['fibonacci', '(', t, '-', 1, ')', '+', 'fibonacci', '(', t, '-', 2, ')'] : [1]));
+      step.rus(this.lime.direct([Math.abs(step.right.value[0])]));
+    });
+    this.algorithms.set('r(arg{rat[1]})', (step) => {
+      step.rus(this.lime.direct([Math.abs(step.right.value[0].n), '/', Math.abs(step.right.value[0].d)]));
     });
   }
 }
@@ -30,4 +29,4 @@ class LimeFunctionFibonacci extends LimeFunction {
 /* ------------------------ division ------------------------ */
 
 // Export
-module.exports = LimeFunctionFibonacci;
+module.exports = LimeFunctionAbsolute;