+++ src/functions/numeric/truncation.js
@@ -12,15 +12,15 @@ class LimeFunctionTruncation extends LimeFunction {
 
     // Operations
     this.operations.r = [
-      'er(arg{int})',
-      'er(arg{rat})',
+      'er(arg{int[1]})',
+      'er(arg{rat[1]})',
     ];
 
     // Algorithms
-    this.algorithms.set('r(arg{int})', (step) => {
+    this.algorithms.set('r(arg{int[1]})', (step) => {
       step.rus(this.lime.direct([step.right.value[0]]));
     });
-    this.algorithms.set('r(arg{rat})', (step) => {
+    this.algorithms.set('r(arg{rat[1]})', (step) => {
       step.rus(this.lime.direct([Math.trunc(step.right.value[0].n / step.right.value[0].d)]));
     });
   }