+++ src/functions/structural/assign.js
@@ -17,7 +17,7 @@ class LimeFunctionAssign extends LimeFunction {
       'tr(_-)',
 
       'cr(var->expr)',
-      'cr(arg{expr}->expr)',
+      'cr(arg{expr[1]}->expr)',
 
       'eb(var,expr)',
     ];